% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force method}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A \code{data.frame} of two non negative numeric columns \code{w} and \code{v}.}

\item{W}{An integer, the knapsack weight limit.}

\item{parallel}{TRUE if parallel search. FALSE by default.}
}
\value{
A list of two elements. The maximum value of the knapsack \code{value} and the elements in the knapsack \code{elements}.
}
\description{
The brute force algorithm will try all possible alternatives of items to include in the knapsack. The algorithm will select the combination of items with the highest value with the maximum possible weight (i.e. less than the Knapsack capacity).
}
\examples{
# Example data
suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
x <- data.frame(
w = sample(1:4000, size = n, replace = TRUE),
v = runif(n = n, 0, 10000))

# Run the algorithm for the first eight values from the example data
brute_force_knapsack(x=x[1:8,], W=3500)

}
\references{
Wikipedia contributors. (2022, May 18). Knapsack problem. In Wikipedia, The Free Encyclopedia. Retrieved 06:38, October 4, 2022, from \url{https://en.wikipedia.org/w/index.php?title=Knapsack_problem&oldid=1088471265}
}
