% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy method}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A \code{data.frame} of two non negative numeric columns \code{w} and \code{v}.}

\item{W}{An integer, the knapsack weight limit.}
}
\value{
A list of two elements.
}
\description{
The greedy algorithm order the elements based on the value per weight unit. It then fills the knapsack until the weight limit is reached.
}
\examples{
# Example data
suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
x <- data.frame(
w = sample(1:4000, size = n, replace = TRUE),
v = runif(n = n, 0, 10000))

# Run the algorithm for the first eight values from the example data
greedy_knapsack(x=x[1:8,], W=3500)

}
\references{
Wikipedia contributors. (2022, May 18). Knapsack problem. In Wikipedia, The Free Encyclopedia. Retrieved 06:38, October 4, 2022, from \url{https://en.wikipedia.org/w/index.php?title=Knapsack_problem&oldid=1088471265}
}
