% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic method}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A \code{data.frame} of two non negative numeric columns \code{w} and \code{v}.}

\item{W}{An integer, the knapsack weight limit.}
}
\value{
A list of two elements.
}
\description{
The dynamic algorithm divides the problem into sub-problems, solves them separately and stores the results in a matrix. It then uses the matrix to solve a wholistic problem by using the stored values when a similar/same problem is encountered.
}
\examples{
# Example data
suppressWarnings(RNGversion(min(as.character(getRversion()),"3.5.3")))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
x <- data.frame(
w = sample(1:4000, size = n, replace = TRUE),
v = runif(n = n, 0, 10000))

# Run the algorithm for the first eight values from the example data
knapsack_dynamic(x=x[1:8,], W=3500)

}
\references{
Wikipedia contributors. (2022, May 18). Knapsack problem. In Wikipedia, The Free Encyclopedia. Retrieved 06:38, October 4, 2022, from \url{https://en.wikipedia.org/w/index.php?title=Knapsack_problem&oldid=1088471265}
}
