\name{knapsackpack-package}
\alias{knapsackpack-package}
\alias{knapsackpack}
\docType{package}
\title{
\packageTitle{knapsackpack}
}
\description{
\packageDescription{knapsackpack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsackpack}
\packageIndices{knapsackpack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{knapsackpack}

Maintainer: \packageMaintainer{knapsackpack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
