% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_disc.R
\docType{package}
\name{KnapsackPack}
\alias{KnapsackPack}
\title{Knapsack algorithm}
\description{
Tree different algorithms to solve the Knapsack problem. Brute force, Dynamic and Greedy.
}
\author{
Leonard Persson Norblad \email{leope744@student.liu.se} Jacob Sibanda \email{jacsi259@student.liu.se}
}
